+++ packages/content-plugin/src/migrations/update-markdown-renderer-feature/compat.ts
@@ -1,7 +1,0 @@
-import { convertNxGenerator } from '@nx/devkit';
-import updateMarkdownRendererFeature from './update-markdown-renderer-feature';
-
-const _default: ReturnType<typeof convertNxGenerator> = convertNxGenerator(
-  updateMarkdownRendererFeature,
-);
-export default _default;