+++ packages/storybook-angular/src/lib/preset.ts
@@ -2,6 +2,7 @@ import { resolve } from 'node:path';
 import { core as PresetCore } from '@storybook/angular/preset';
 import { fileURLToPath } from 'node:url';
 import type { Plugin, UserConfig } from 'vite';
+import * as vite from 'vite';
 
 export const previewAnnotations = async (
   entries: string[] = [],
@@ -111,7 +112,7 @@ export const viteFinal = async (config: any, options: any): Promise<any> => {
             : 'css',
       }),
       angularOptionsPlugin(options, { normalizePath, experimentalZoneless }),
-      storybookEsbuildPlugin(),
+      storybookTransformConfigPlugin(),
     ],
     define: {
       STORYBOOK_ANGULAR_OPTIONS: JSON.stringify({
@@ -208,13 +209,13 @@ function angularOptionsPlugin(
   };
 }
 
-function storybookEsbuildPlugin(): Plugin {
+function storybookTransformConfigPlugin(): Plugin {
   return {
-    name: 'analogjs-storybook-esbuild-config',
+    name: 'analogjs-storybook-transform-config',
     apply: 'build',
     config() {
       return {
-        esbuild: {
+        [vite.rolldownVersion ? 'oxc' : 'esbuild']: {
           // Don't mangle class names during the build
           // This fixes display of compodoc argtypes
           keepNames: true,